requirements-test.txt
pytest
pytest-httpx
fakeredis
responses
prometheus-client
